FIRE-33958 - Tentative fix for zombie notifications
I've only had limited success in reproducing this, so that fact I can't repro after fixing is not that compelling. However, when the issue occurs the underlying notification has been removed but the UI remains, this seems consistent with notification list being updated outside of the mutex guard and causing issues.
Note: LLNotifications has a number of methods that do an initial find and then call updateItem. I have not, at this point locked the higher livel calls as the updateItem call itself is locked and the higher level find is not likely to be problematic. If problems persist we can try guarding those too.