Modify Firestorm.app embedded path.
Based on `http://qin.laya.com/tech_coding_help/dylib_linking.html` (thanks Brad)
In the past, I tried this, but I think it failed to find the executable. 
Given that the change needs to be made to the executable, this should have worked. Trying again, but this time with Contents folder in the right place.