viewer#2363 Region Day Offset error with Time of Day
When we have a 23h day cycle and 23h offset, doing a 24h flip will
result in -1h. Flip by day length instead so that 23h-23h will result
in 0h.

This has a negative side effect of values not being saved exactly as
users specified them, so a -1h offset will turn to 4h on next load
with a 5h long day. But it is nature of a day to 'loop', so I decided
there is no point to translate -1 into 14 or something like that.