FIRE-33989 - Noto-Sans font fallbacks not finding the default
Applying a correction to replace the non-existent NotoSans.ttf as the default fallback. 

It has been "wrong" for a long time suggesting that it was an unrelated change (possibly aligning our fallbacks with LL) that triggered this, so there may be other issues.

We may need to review this further if it has other side-effects.