Add colour to the FPS Text, relocate vsync next to fps limit
if there are no limits on FPS it will remain white.
if vsync or fps limit are used and the fps is below the lower of the two limits then it will be coloured yellow, if we are at or above a limit then it will be green.
Clicking the fps text opens the preferences panel with limit framerate highlighted.