FIRE-35081 - Blurry prims not changing with graphics settings, not happening with SL Viewer
Fixed the blurry texture issue with a couple of fixes.
Objects selected by by the user would gain a BOOST_SELECTED boost level, but when unselected the system did not store the old value and  would get turned to BOOST_NONE which allows the Bias system to further downscale the image lower then the original size.
Upon first loading a texture, if the desired size was lower then what was decoded, it would auto scale down the new texture, without first validating if the texture was allowed to be scaled down due to type of texture.
Further, the rules applied to the LLViewerTextureList::updateImageDecodePriority would apply to all textures on screen no matter what type as long as the boost was lower then BOOST_HIGH.

Changes include new mPrevBoostLevel for any texture with accessor methods to restore the values and updated the code which first set's the boost level to BOOST_SELECTED to store the boost level of all affected objects before changing the boost level.

Added LL's "Fix" for the bias by forcing to to stay static if VRAM or System RAM is at 10% of max.

Added logic to the updateImageDecodePriority to not apply bias to non LOD textures as well as any texture which has a current discard of 0 or 1, so not to blur high res images close to the screen.

Added logic to updateImagesCreateTextures to only scale down LOD textures which have a boost level < BOOST_HIGH

Added sPrevDesiredDiscardBiasto keep track other previous bias value before it was updated to allow for better control of the system handling bias effects such if the bias value is not chanigng, use the normal number of texture updates.

Fixed up the Developer -> Consoles -> Texture Console to have correct values for the
llformat("Tex: %d Fetch: %d(%d) Pkts:%d(%d) CAC R/W: %d/%d LFS:%d RAW:%d HTP:%d DEC:%d CRE:%d FCA:%d ",
line.