Fix issue with DebugShowColor value
In the LLViewerWindow.cpp update method, the DebugShowColor was loaded from LLCachedControl as a Bool, when in fact it is stored as a S32 value.
This causes a exception to trigger for the  load code in RelWithDebInfo mode and may cause issues in Release.
Changed the code to treat as S32 and check the value to == 1 as it is only goes from 0 to 1