FIRE-35508 - Float text background change with Highlight disabled
User requested that UI allow background for floating text to be modified if the highlight check box is disabled.

Removed the UI XML enabled_control requirement.

Only issue is if drop down option "Only Highlighted Objects" does nothing if highlight option is disabled as it is dependent on the check to be performed.