FIRE-34977 - additional improvement WMI query cache
In testing the WMI device driver query could take up to 50ms, it might be more for devices with multiple cards and or slower hardware. Given that the device is not realistically going to change during a run we will cache the results. Worst case it will be wrong until the next restart.