#4794 SceneLoadRearMaxRadiusFraction not stored as a fraction - <FS:minerjr> Cherry-pick to fix issue FIRE-36003
Reported issue to LL GitHub and FS JIRA FIRE-36003
LL Issue: https://github.com/secondlife/viewer/issues/4794
Linked to FS issue and cherry picked to pull in fix.

Inital Issue:
When you look at the SceneLoadRearMaxRadiusFraction value stored in the settings.xml, you will notice that the value is set to 75.0 for a float value. And the comments mention it is suppose to be fraction.

The code that uses it is for scaling the amount of the scene behind the visible camera is loaded. The issue is because the calculation no longer is divided by 100.0f, the draw distance is multiplied by the value and then clamped at the draw distance value. So it is effectively only ever the draw distance and not the default 75% of it. So a 256 draw distance is locked at 256 meters instead of 192.

This causes more objects to load behind the user, especially very far objects which can cause more RAM and VRAM usage, especially on teleport.